immersatable plugin improvements
