+++ superset-frontend/packages/superset-plugin-chart-immersatable/package.json
@@ -49,7 +49,8 @@
     "react-window": "^1.8.8",
     "@visx/vendor": "3.5.0",
     "moment": "^2.25.3",
-    "memoize-one": "^5.1.1"
+    "memoize-one": "^5.1.1",
+    "@react-icons/all-files": "^4.1.0"
   },
   "peerDependencies": {
     "@superset-ui/chart-controls": "*",
