+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/components/AreaChart/index.ts
@@ -1,0 +1,1 @@
+export * from './AreaChart';
