+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/components/DataTable/index.tsx
@@ -1,0 +1,1 @@
+export * from './DataTable';
