+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/components/LineSeriesChart/index.ts
@@ -1,0 +1,2 @@
+export * from './LineSeriesChart';
+export * from './TooltipContent';
