+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './DataTable';
+export * from './LineSeriesChart';
+export * from './AreaChart';
