+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/utils.ts
@@ -1,14 +1,5 @@
 import { format, parseISO } from 'date-fns';
 import moment, { Moment } from 'moment';
-import {
-  DataRecordValue,
-  isProbablyHTML,
-  sanitizeHtml,
-  GenericDataType,
-  getNumberFormatter,
-} from '@superset-ui/core';
-import { DataColumnMeta } from './plugin/transformProps';
-import DateWithFormatter from './DateWithFormatter';
 
 export const ChartAxisLabelColor = '#111827';
 
@@ -263,105 +254,3 @@ export function getSinceUntil(
 
   return [_since, _until];
 }
-
-function formatValue(
-  formatter: DataColumnMeta['formatter'],
-  value: DataRecordValue,
-): [boolean, string] {
-  // render undefined as empty string
-  if (value === undefined) {
-    return [false, ''];
-  }
-  // render null as `N/A`
-  if (
-    value === null ||
-    // null values in temporal columns are wrapped in a Date object, so make sure we
-    // handle them here too
-    (value instanceof DateWithFormatter && value.input === null)
-  ) {
-    return [false, 'N/A'];
-  }
-  if (formatter) {
-    return [false, formatter(value as number)];
-  }
-  if (typeof value === 'string') {
-    return isProbablyHTML(value) ? [true, sanitizeHtml(value)] : [false, value];
-  }
-  return [false, value.toString()];
-}
-
-export function formatColumnValue(
-  column: DataColumnMeta,
-  value: DataRecordValue,
-) {
-  const { dataType, formatter, config = {} } = column;
-  const isNumber = dataType === GenericDataType.Numeric;
-  const smallNumberFormatter =
-    config.d3SmallNumberFormat === undefined
-      ? formatter
-      : getNumberFormatter(config.d3SmallNumberFormat);
-  return formatValue(
-    isNumber && typeof value === 'number' && Math.abs(value) < 1
-      ? smallNumberFormatter
-      : formatter,
-    value,
-  );
-}
-
-export const checkChartData = (cellValue: string) => {
-  if (
-    cellValue?.toString().includes('[') &&
-    Array.isArray(JSON.parse(cellValue as string))
-  ) {
-    return true;
-  }
-  return false;
-};
-
-interface CustomData {
-  [key: string]: any;
-}
-
-const filterDataByDateInterval = (
-  data: string,
-  startDate: Date,
-  endDate: Date,
-) =>
-  JSON.parse(data).filter((item: (string | number | Date)[]) => {
-    const itemDate = new Date(item[0]);
-    return itemDate >= startDate && itemDate <= endDate;
-  });
-
-export const processCustomData = (
-  myData: CustomData[],
-  timeRangeCols: string[],
-  timeSinceUntil: {
-    startDate: Date;
-    endDate: Date;
-  },
-) => {
-  const { startDate, endDate } = timeSinceUntil;
-  return myData.map(customdata => {
-    const updatedCustomdata = { ...customdata };
-    const keysArray = Object.keys(updatedCustomdata);
-    // eslint-disable-next-line no-restricted-syntax
-    for (const item of keysArray) {
-      const value = updatedCustomdata[item];
-      if (
-        timeRangeCols.includes(item) &&
-        startDate &&
-        endDate &&
-        value?.toString().includes('[') &&
-        Array.isArray(JSON.parse(value))
-      ) {
-        const filteredData = filterDataByDateInterval(
-          value,
-          startDate,
-          endDate,
-        );
-        updatedCustomdata[item] = JSON.stringify(filteredData);
-      }
-    }
-    return updatedCustomdata;
-  });
-};
