+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/utils/checkChartData.ts
@@ -1,0 +1,9 @@
+export const checkChartData = (cellValue: string) => {
+  if (
+    cellValue?.toString().includes('[') &&
+    Array.isArray(JSON.parse(cellValue as string))
+  ) {
+    return true;
+  }
+  return false;
+};
